Functions that depend on non-path-params are now re-run as expected
Implemented by caching the function + args in temp dir. Surely there's
a better way, but bringing in a database felt dirty and opaque (how do you
clear the fact that they've run?)

Note: Perhaps wouldn't matter if DB was opaque if tasks had a .clean and/or
a _force_run argument. Hmm.