Improve tracebacks in Py2.
Turns out, raise_from does not imitate PY3's traceback behavior.
Included a hack to basically attach more information to an Exception
as it bubbles up, since it's not clear from the tracebacks upon first
entry to a task what exactly happened. Could probably use this within
invoke as well to explain what task we were trying to run, etc.
But please move to Python3 instead of using this wrapper :)